requirements.txt
Django==2.1
djangorestframework==3.9.1
coreapi==2.3.3
django-rest-swagger==2.2.0


